Bluesky blew through max int number of messages
Well, end of an era I guess. Time to use a bigger (bigint) cursor.

Interestingly enough, some of Rust parts were already using an i64,
but others weren't. Huh.

+++ src/services/bluesky/client.rs
@@ -145,7 +145,7 @@ impl Bluesky {
     pub async fn subscribe_to_operations<P: CommitProcessor>(
         &self,
         processor: &P,
-        cursor: Option<i32>,
+        cursor: Option<i64>,
     ) -> Result<()> {
         let url = match cursor {
             Some(cursor) => format!(
